Refactor: Changed Pickups.channel and Game.channel to TextChannels
