+++ commands/pickups/add.js
@@ -1,5 +1,3 @@
-const { readyHandler } = require('../../libs/handlers');
-
 const { MessageEmbed } = require('discord.js');
 
 const run = async(message) => {
@@ -13,25 +11,23 @@ const run = async(message) => {
 
     if (!pickupsNames) return message.reply('No pickups found!');
 
-    const pugger = await message.client.puggers.fetch(message.author.id);
+    const pugger = await message.client.puggers.fetch(message.author);
     if (!pugger) return message.reply('Couldn\'t resolve user!');
 
     const joined = new Array();
 
     if (pickupsNames instanceof Array) {
         for (const pickupName of pickupsNames) {
-            const res = await pugger.queue(message.channel.id, pickupName);
-            if (res) {
-                if (res.isFull) readyHandler(res.game, message.channel);
+            const res = await pugger.queue(message.channel, pickupName);
+            if (res)
                 joined.push(res.game);
-            }
         }
     } else {
-        const channel = await message.client.pickups.fetchChannel(message.channel.id);
+        const channel = await message.client.pickups.fetchChannel(message.channel);
         for (const { name } of channel) {
             const res = await pugger.queue(message.channel.id, name);
             if (res) {
-                if (res.isFull) readyHandler(res.game, message.channel);
+                if (res.isFull) res.game.readyHandler();
                 joined.push(res.game);
             }
         }
